+++ src/prodbg/backend/backend_requests.h
@@ -73,7 +73,7 @@ public:
     // Requests a block of memory from the target. The return vector has
     // is uint16_t with the real data in lower 8-bit and upper is reserved for
     // status flags Readable/Writeable/etc
-    bool request_memory(uint64_t lo, uint64_t hi, QVector<uint16_t>* target);
+    bool request_memory(uint64_t lo, uint64_t hi, QVector<uint8_t>* target);
 
 private:
     Q_SIGNAL void request_basic_signal(IBackendRequests::BasicRequest request_type);
@@ -82,7 +82,7 @@ private:
     Q_SIGNAL void request_remove_file_line_breakpoint_signal(const QString& filename, int line);
     Q_SIGNAL void request_locals_signal(const IBackendRequests::ExpandVars& expanded_vars, uint64_t request_id);
     Q_SIGNAL void request_frame_index_signal(int frame_index);
-    Q_SIGNAL void request_memory_signal(uint64_t lo, uint64_t hi, QVector<uint16_t>* target);
+    Q_SIGNAL void request_memory_signal(uint64_t lo, uint64_t hi, QVector<uint8_t>* target);
 
     /*
     Q_SIGNAL void evalExpression(const QString& expr, uint64_t* out);
